Ratings not exists but a bug exists.
If you post a rating before posting a comment the site will not show because it is expecting a to already be there first.

The program is searching for a value that dosent exists and this causes the bug. Everything will still be stored in the database, will fix this later.